Command line scrabble in C
The first player combines two or more of his or her letters to form a word,
and places the word on the board to read either across or down with one letter
on the center square. Diagonal words are not allowed.

Words are placed by specifying the row and column. When the word is completely placed,
the player's score will be posted in the window to the right.

Signed-off-by: Javeria Khan <javeria.ak@gmail.com>